libkpod -> lib rename
