+++ lib/hooks.go
@@ -1,4 +1,4 @@
-package libkpod
+package lib
 
 import (
 	"encoding/json"
