+++ lib/kill.go
@@ -1,4 +1,4 @@
-package libkpod
+package lib
 
 import (
 	"github.com/docker/docker/pkg/signal"
